requirements.txt
streamlit
google-generativeai
youtube-transcript-api==1.2.2
langchain
langchain-google-genai
langchain-chroma
python-dotenv
google-genai
gtts
langchain
langchain_community
langchain-google-genai
streamlit
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
pypdf
rapidocr-onnxruntime
requests